26 message bus communication

+++ ForumService/ForumServiceLogic/ForumLogic.cs
@@ -1,4 +1,5 @@
 ﻿using ForumServiceDAL;
+using ForumServiceMessageBusProducer;
 using ForumServiceModels;
 using ForumServiceModels.Interfaces;
 using Microsoft.Extensions.Logging;
@@ -9,14 +10,16 @@ public class ForumLogic : IForumLogic
 {
     private readonly IForumRepository _repository;
     private readonly ILogger<IForumLogic> _logger;
-    public ForumLogic(ILogger<IForumLogic> logger, IForumRepository repository)
+    private readonly ForumMessageBusProducer _producer;
+    public ForumLogic(ILogger<ForumLogic> logger, IForumRepository repository, ForumMessageBusProducer producer)
     {
         _logger = logger;
         _repository = repository;
+        _producer = producer;
     }
     public Forum? GetForum(string name)
     {
-        _logger.Log(LogLevel.Information, "Getting forum with name {id}", name);
+        _logger.Log(LogLevel.Information, "Getting forum with name {Name}", name);
         return _repository.GetForum(name);
     }
 
@@ -30,12 +33,14 @@ public class ForumLogic : IForumLogic
     {
         if (_repository.ForumExists(forum.Name))
         {
-            _logger.Log(LogLevel.Information, "Forum with name {name} already exists", forum.Name);
+            _logger.Log(LogLevel.Information, "Forum with name {Name} already exists", forum.Name);
             return null;
         }
-        _logger.Log(LogLevel.Information, "Adding forum {forum}", forum);
+        _logger.Log(LogLevel.Information, "Adding forum {Forum}", forum.Name);
         if (_repository.AddForum(forum))
         {
+            ForumShared forumShared = new() { Name = forum.Name, AdminId = forum.AdminId};
+            _producer.SendForumCreatedMessage(forumShared);
             return _repository.GetForum(forum.Name);
         }
         return null;
@@ -43,13 +48,40 @@ public class ForumLogic : IForumLogic
 
     public Forum? UpdateForum(Forum forum)
     {
-        _logger.Log(LogLevel.Information, "Updating forum {forum}", forum);
+        _logger.Log(LogLevel.Information, "Updating forum {Forum}", forum.Name);
         return _repository.UpdateForum(forum) ? _repository.GetForum(forum.Name) : null;
     }
 
-    public bool DeleteForum(string name)
-    { 
-        _logger.Log(LogLevel.Information, "Deleting forum with name {name}", name);
+    public bool DeleteForum(string name, int id)
+    {
+        Forum? forum = _repository.GetForum(name);
+        if (forum == null)
+        {
+            _logger.LogInformation("Attempted to delete forum that does not exist: {Name}", name);
+            return false;
+        }
+
+        if (forum.AdminId != id)
+        {
+            _logger.LogInformation("User with id {Id} attempted to delete forum {Name} but is not admin", id, name);
+            return false;
+        }
+
         return _repository.DeleteForum(name);
+    }
+
+    public bool AddAccount(Account account)
+    {
+        return _repository.AddAccount(account);
+    }
+
+    public bool DeleteAccount(Account account)
+    {
+        return _repository.DeleteAccount(account);
+    }
+
+    public int GetAccountIdFromGoogleId(string googleId)
+    {
+        return _repository.GetAccountIdFromGoogleId(googleId) ?? -1;
     }
 }
